requirements.txt
psycopg2-binary
xlrd
openpyxl
python-dotenv
fastapi
uvicorn[standard]
google-genai